+++ lib/lite/validators/ip_address_validator.rb
@@ -20,11 +20,30 @@ class IpAddressValidator < BaseValidator
     assert_valid_option!(:protocol, REGEXP.keys.push(:any))
   end
 
+  def error_message_for(option)
+    options[:message] || I18n.t("errors.messages.ip_address.#{option}")
+  end
+
   def protocol
     options[:protocol] || :any
   end
 
-  def valid_attr?
+  def valid?
+    valid_format? && valid_address?
+  end
+
+  def valid_address?
+    addresses = options[:include_address] || options[:exclude_address]
+    return true unless addresses
+
+    check = options[:include_address] ? :any? : :none?
+    check = Array(addresses).send(check) { |address| value.include?(address.to_s) }
+    return true if check
+
+    record.errors.add(attribute, error_message_for(:address))
+  end
+
+  def valid_format?
     case protocol
     when :ipv4 then valid_regexp?(:ipv4)
     when :ipv6 then valid_regexp?(:ipv6)